Moved chart string parsing into chart parse assembly
