+++ PumpItUpUCSImporter/UCSImporter.Data/AndamiroChartStepInfoRepository.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.Logging;
 using UCSImporter.Data.Apis.Contracts;
-using UCSImporter.Domain.ChartParser.Enums;
+using UCSImporter.Domain.ChartParser.Exceptions;
 using UCSImporter.Domain.ChartParser.Models;
 using UCSImporter.Domain.Contracts;
 using UCSImporter.Domain.ValueTypes;
@@ -26,85 +26,18 @@ public sealed class AndamiroChartStepInfoRepository : IChartStepInfoRepository
         foreach (var id in chartIds)
         {
             var infoString = await _api.GetChartFile(id, cancellationToken);
-            var rows = ExtractRows(id, infoString);
-            result.Add(new ChartStepInfo(id, rows));
+            try
+            {
+                var info = ChartParser.ParseAndamiroChart(id, infoString);
+                result.Add(info);
+            }
+            catch (ChartRowsInvalidException ex)
+            {
+                result.Add(ex.ParsedInfo);
+                _logger.LogError(ex.Message, ex);
+            }
         }
 
         return result;
-    }
-
-    private IEnumerable<ChartRow> ExtractRows(ChartId chartId, string chart)
-    {
-        var rows = new List<ChartRow>();
-
-        var chartState = new ChartState(0, 0, 0, 0);
-        foreach (var rowString in chart.Replace("\r", "").Split("\n"))
-            try
-            {
-                if (rowString.StartsWith(":Format", StringComparison.OrdinalIgnoreCase)
-                    || rowString.StartsWith(":Mode", StringComparison.OrdinalIgnoreCase))
-                    continue;
-
-                if (rowString.StartsWith(":BPM=", StringComparison.OrdinalIgnoreCase))
-                {
-                    chartState = chartState.ApplyNewBpm(double.Parse(rowString.Replace(":BPM=", "")));
-                    continue;
-                }
-
-                if (rowString.StartsWith(":Delay=", StringComparison.OrdinalIgnoreCase))
-                {
-                    chartState = chartState.ApplyNewDelay(double.Parse(rowString.Replace(":Delay=", "")));
-                    continue;
-                }
-
-                if (rowString.StartsWith(":Beat=", StringComparison.OrdinalIgnoreCase))
-                {
-                    chartState = chartState.ApplyNewBeat(double.Parse(rowString.Replace(":Beat=", "")));
-                    continue;
-                }
-
-                if (rowString.StartsWith(":Split=", StringComparison.OrdinalIgnoreCase))
-                {
-                    chartState = chartState.ApplyNewSplit(double.Parse(rowString.Replace(":Split=", "")));
-                    continue;
-                }
-
-                var leftPad = new Pad(StateFromCharacter(rowString[0]), StateFromCharacter(rowString[1]),
-                    StateFromCharacter(rowString[2]),
-                    StateFromCharacter(rowString[3]),
-                    StateFromCharacter(rowString[4]));
-                var rightPad = rowString.Length == 10
-                    ? new Pad(
-                        StateFromCharacter(rowString[5]),
-                        StateFromCharacter(rowString[6]),
-                        StateFromCharacter(rowString[7]),
-                        StateFromCharacter(rowString[8]),
-                        StateFromCharacter(rowString[9]))
-                    : new Pad(ArrowState.Nothing, ArrowState.Nothing, ArrowState.Nothing, ArrowState.Nothing,
-                        ArrowState.Nothing);
-                var newRow = new ChartRow(leftPad, rightPad, chartState);
-
-                if (rows.Count > 0) newRow.SetPreviousRow(rows.Last());
-
-                rows.Add(newRow);
-            }
-            catch (Exception e)
-            {
-                _logger.LogWarning($"Row {rowString} couldn't be parsed for chart {chartId}", e);
-            }
-
-        return rows;
-    }
-
-    private static ArrowState StateFromCharacter(char character)
-    {
-        switch (character)
-        {
-            case '.': return ArrowState.Nothing;
-            case 'X': return ArrowState.Step;
-            case 'M': return ArrowState.StartOfFreeze;
-            case 'W': return ArrowState.EndOfFreeze;
-            default: throw new Exception($"Unknown arrow state {character}");
-        }
     }
 }
