+++ PumpItUpUCSImporter/UCSImporter.Domain.ChartParser/Models/RowParseError.cs
@@ -1,0 +1,13 @@
+﻿namespace UCSImporter.Domain.ChartParser.Models;
+
+internal class RowParseError
+{
+    public RowParseError(string rowString, Exception exception)
+    {
+        RowString = rowString;
+        Exception = exception;
+    }
+
+    public string RowString { get; }
+    public Exception Exception { get; }
+}
