+++ PumpItUpUCSImporter/UCSImporter.Domain.ChartParser/UCSImporter.Domain.ChartParser.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
